Refactor one-time-token stuff a bit, and move stuff into service.
Also we're returning the expiry date and ttl now.
